refactor(#46): BaseEntity 코드 변경 및 폴더 수정
